Move sqlalchemy utils functions
